requirements.txt
Cython==3.0.6
gdown==4.7.1
GitPython==3.1.40
hickle==5.0.2
matplotlib
numpy
openai
opencv_python==4.8.1.78
Pillow==10.1.0
progressbar33==2.4
pybullet==3.2.6
PySide6==6.5.2
scipy
shapely==2.0.2
tqdm==4.66.1
